make cypress/react -> react 18, cypress/react17 -> older versions
